[update] Friend: friend.htmlを追加 & setupFriendEventListenersの呼び出しを調整し、Friend申請APIの重複呼び出しを修正
