+++ Makefile
@@ -93,14 +93,15 @@ start:
 
 .PHONY: down
 down:
-	COMPOSE_PROFILES=elk,blockchain,monitor,three docker-compose $(COMPOSE_FILES_ARGS) down; \
-	PATTERN='127.0.0.1 $(SERVER_NAME)'; \
-	OSTYPE=`uname -s`; \
-	if [ "$$OSTYPE" = "Darwin" ]; then \
-		sudo sed -i '' "/$$PATTERN/d" /etc/hosts; \
-	elif [ "$$OSTYPE" = "Linux" ]; then \
-		sudo sed -i "/$$PATTERN/d" /etc/hosts; \
-	fi 
+	COMPOSE_PROFILES=elk,blockchain,monitor,three docker-compose $(COMPOSE_FILES_ARGS) down
+#	COMPOSE_PROFILES=elk,blockchain,monitor,three docker-compose $(COMPOSE_FILES_ARGS) down; \
+#	PATTERN='127.0.0.1 $(SERVER_NAME)'; \
+#	OSTYPE=`uname -s`; \
+#	if [ "$$OSTYPE" = "Darwin" ]; then \
+#		sudo sed -i '' "/$$PATTERN/d" /etc/hosts; \
+#	elif [ "$$OSTYPE" = "Linux" ]; then \
+#		sudo sed -i "/$$PATTERN/d" /etc/hosts; \
+#	fi
 
 .PHONY: d
 d:
@@ -183,7 +184,7 @@ fclean: down docker_rm remove_mount_volume
 ## build初期設定: .envの作成 & /etc/hostsにserver_nameを登録
 .PHONY: init
 init: env cert_key
-	@chmod +x init/add_host.sh && ./init/add_host.sh init/.os_env
+	@#chmod +x init/add_host.sh && ./init/add_host.sh init/.os_env
 	@chmod +x init/make_dir.sh && ./init/make_dir.sh init/.os_env
 
 .PHONY: env
