+++ docker/srcs/uwsgi-django/accounts/templates/accounts/friends.html
@@ -1,0 +1,24 @@
+<!-- user.html -->
+
+{% load static %}
+
+<!-- <div class="canvas-container"> -->
+<!-- <div class="hs-text"> -->
+<div class="container" >
+
+    <div class="friends-info-container">
+        <ul id="friends-container">
+            <!-- JavaScriptで動的に内容が生成される -->
+        </ul>
+        <hr>
+        <ul id="requests-container">
+            <!-- JavaScriptで動的に内容が生成される -->
+        </ul>
+    </div>
+
+</div>
+<!-- </div> -->
+<!-- </div> -->
+
+<!--<script type="module" src="{% static 'accounts/js/friend.js' %}"></script>-->
+<!--<script type="module" src="{% static 'accounts/js/online-status.js' %}"></script>-->
