+++ docker/srcs/uwsgi-django/accounts/templates/accounts/user.html
@@ -11,20 +11,10 @@
         <ul id="user-info">
             <!-- JavaScriptで動的に内容が生成される -->
         </ul>
-        <hr>
-        <ul id="friends-container">
-            <!-- JavaScriptで動的に内容が生成される -->
-        </ul>
-        <hr>
-        <ul id="requests-container">
-            <!-- JavaScriptで動的に内容が生成される -->
-        </ul>
     </div>
 
 </div>
 <!-- </div> -->
 <!-- </div> -->
 
-<script type="module" src="{% static 'accounts/js/userProfile.js' %}"></script>
-<script type="module" src="{% static 'accounts/js/friend.js' %}"></script>
-<script type="module" src="{% static 'accounts/js/online-status.js' %}"></script>
+<!--<script type="module" src="{% static 'accounts/js/userProfile.js' %}"></script>-->
