+++ docker/srcs/uwsgi-django/accounts/urls.py
@@ -7,6 +7,7 @@ from accounts.views.basic_auth import LogoutTemplateView
 from accounts.views.basic_auth import LoginTemplateView
 from accounts.views.user import UserProfileView, get_user_info, ChangeAvatarView
 from accounts.views.user import EditUserProfileTemplateView
+from accounts.views.friend import UserFriendsView
 from accounts.views.oauth import OAuthWith42
 from accounts.views.two_factor_auth import Disable2FaView
 from accounts.views.two_factor_auth import Enable2FaTemplateView
@@ -27,6 +28,8 @@ urlpatterns = [
     path('user/', UserProfileView.as_view(), name='user'),
     path('edit/', EditUserProfileTemplateView.as_view(), name='edit'),
     path('change-avatar/', ChangeAvatarView.as_view(), name='change_avatar'),
+
+    path('friends/', UserFriendsView.as_view(), name='friends'),
 
     path('info/<str:nickname>/', get_user_info, name='info'),
