+++ docker/srcs/uwsgi-django/trans_pj/static/spa/js/routing/routeTable.js
@@ -10,7 +10,7 @@ import GameHistory  from "../views/UserProfile.js";
 
 import UserProfile  from "../views/UserProfile.js";
 import UserInfo     from "../views/UserInfo.js";
-import Friend       from "../views/Friend.js";
+import Friends      from "../views/Friends.js";
 import DMSessions   from "../views/DMSessions.js";
 import DMwithUser   from "../views/DMwithUser.js";
 
@@ -48,7 +48,7 @@ export const routeTable = [
 
   { path: "/user-profile/",         view: UserProfile },  // private profile
   { path: "/user-info/:nickname/",  view: UserInfo },     // public profile
-  { path: "/user-friend/",          view: Friend },
+  { path: "/user-friends/",         view: Friends },
   { path: "/dm/",                   view: DMSessions },
   { path: "/dm-with/:nickname/",    view: DMwithUser },
