--- docker/srcs/uwsgi-django/trans_pj/static/spa/js/views/Friend.js
@@ -1,20 +1,0 @@
-import AbstractView from "./AbstractView.js";
-import fetchData from "../utility/fetch.js";
-import { getUrl } from "../utility/url.js";
-import { loadAndExecuteScript } from "../utility/script.js";
-
-export default class extends AbstractView {
-  constructor(params) {
-    super(params);
-    this.setTitle("Friend");
-  }
-
-  async getHtml() {
-    const uri = "/accounts/user/";
-    const data = await fetchData(uri);
-    return data;
-  }
-  async executeScript() {
-    //executeScriptTab("");
-  }
-}
