+++ docker/srcs/uwsgi-django/trans_pj/templates/header.html
@@ -83,7 +83,7 @@
 			<li class="nav-item"><a class="nav-link hth-nav-link hth-menu-fsize nav__link"
 									href="/user-profile/" data-link>User-Profile</a></li>
 			<li class="nav-item"><a class="nav-link hth-nav-link hth-menu-fsize nav__link"
-									href="/user-friend/" data-link>Friend</a></li>
+									href="/user-friends/" data-link>Friend</a></li>
 			<li class="nav-item"><a class="nav-link hth-nav-link hth-menu-fsize nav__link"
 									href="/dm/" data-link>DM</a></li>
